feat: implement security interruption and modernize C++ codebase
- Implement interactive interruption for [SECURITY] issues in CLI and Web/WASM.
- CLI uses TTY-aware std::cin prompt; Web uses window.confirm via Emscripten.
- Instrument checkers (Archive, Binary, ModelDescription, Schema, Directory) and ModelChecker to respect abort flag.
- Stabilize clang-tidy environment using clang-tidy-19 and libc++.
- Perform extensive codebase modernization: fix hundreds of const-correctness, bracing, and [[nodiscard]] warnings.
- Update AGENTS.md with stabilized environment instructions.
- Ensure all 49 unit tests pass and web build/linting is clean.

+++ core/checker/archive/archive_checker.cpp
@@ -192,7 +192,8 @@ void ArchiveChecker::checkZipBomb(const std::vector<ZipFileEntry>& entries, Cert
         // Check compression ratio
         if (entry.compressed_size > 0)
         {
-            const double ratio = static_cast<double>(entry.uncompressed_size) / entry.compressed_size;
+            const double ratio =
+                static_cast<double>(entry.uncompressed_size) / static_cast<double>(entry.compressed_size);
             if (ratio > MAX_RATIO)
             {
                 test.status = TestStatus::FAIL;
